Add Message class for email.
