+++ tests/test_app/TestApp/Mailer/TestEmail.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+namespace TestApp\Mailer;
+
+use Cake\Mailer\Email;
+
+/**
+ * Help to test Email
+ */
+class TestEmail extends Email
+{
+    protected $messageClass = TestMessage::class;
+}
